refactor: split instagram-poster into modular lib/ structure + fix code review issues
Split the 1081-line post.mjs monolith into 8 focused modules under lib/:
- fetch.mjs (shared fetchJson, pick, IG_GRAPH — also used by analytics)
- history.mjs (post history persistence)
- art-fetchers.mjs (Harvard/Met/AIC random, specific, seasonal)
- dropbox.mjs (upload, share link, cleanup)
- token-refresh.mjs (Meta token auto-refresh)
- instagram-api.mjs (Graph API publish, reels, auto-story)
- captions.mjs (caption, hashtags, alt text builders)
- render.mjs (moved from root, watercolor card renderer)

Bug fixes from code review:
- Extract shared Dropbox upload helper (was duplicated in uploadImage + publishReel)
- Add res.ok check in waitForContainer (was spinning on auth failures)
- Fix array mutation in source shuffling
- Remove phantom 'rijks' from ArtSourceId union (no adapter exists)
- Add env var guard in HarvardAdapter
- Use Promise.allSettled in fetchById without source
- Add console.warn in analytics getMediaInsights catch
- Parallelize serial await in analytics loop with Promise.all
- Replace magic number 99 with SEASONAL_COOLDOWN_RESET constant
- Extract shared renderCard with CARD_PRESETS in render.mjs
- Handle null canvas context in storyCardRenderer.ts
- Use Met API dedicated filter params for medium and century
